rename upload_data -> s3_management
