+++ collector/spot-dataset/gcp/lambda/s3_management.py
